BookoAPI/requirements.txt
python = 3.9
django = 3.2
djangorestframework = 3.12.4